pyproject.toml
[tool.poetry]
name = "psychology"
version = "0.1.0"
description = ""
authors = ["Almagro Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.214.0"}
pydantic = "1.10.2"
django-cors-headers = "^4.3.1"
psycopg2-binary = "*"


# [tool.poetry.group.database.dependencies]
# dj-database-url = "^2.1.0"
# psycopg2 = "^2.9.9"

[tool.poetry.group.server.dependencies]
uvicorn = "^0.24.0.post1"
httpx = "^0.25.1"

[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"

[tool.poetry.group.lint.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
safety = "^2.3.5"
flake8-docstrings = "^1.7.0"
flake8-import-order = "^0.18.2"
flake8-bugbear = "^23.9.16"
pep8-naming = "^0.13.3"
flake8-bandit = "^4.1.1"
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
  | migrations
)/poetry config http-basic.

'''

[tool.coverage.run]
omit = [
    "apps/*/migrations/*",
]
